+++ examples/suggest-cli/src/main.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use std::sync::Arc;
+use std::{collections::HashMap, sync::Arc};
 
 use anyhow::Result;
 use clap::{Parser, Subcommand, ValueEnum};
@@ -50,7 +50,8 @@ enum Commands {
     },
     /// Query against ingested data
     Query {
-        provider: SuggestionProviderArg,
+        #[clap(long, short)]
+        provider: Option<SuggestionProviderArg>,
         /// Input to search
         input: String,
     },
@@ -101,8 +102,8 @@ fn main() -> Result<()> {
         Commands::Ingest {
             reingest,
             providers,
-        } => ingest(&store, reingest, providers),
-        Commands::Query { provider, input } => query(&store, provider, input),
+        } => ingest(&store, reingest, providers, cli.verbose),
+        Commands::Query { provider, input } => query(&store, provider, input, cli.verbose),
     };
     Ok(())
 }
@@ -125,12 +126,17 @@ fn build_store(cli: &Cli) -> Arc<SuggestStore> {
         .unwrap_or_else(|e| panic!("Error building store: {e}"))
 }
 
-fn ingest(store: &SuggestStore, reingest: bool, providers: Vec<SuggestionProviderArg>) {
+fn ingest(
+    store: &SuggestStore,
+    reingest: bool,
+    providers: Vec<SuggestionProviderArg>,
+    verbose: bool,
+) {
     if reingest {
-        println!("Reingesting data...");
+        print_header("Reingesting data");
         store.force_reingest();
     } else {
-        println!("Ingesting data...");
+        print_header("Ingesting data");
     }
     let constraints = if providers.is_empty() {
         SuggestIngestionConstraints::all_providers()
@@ -141,26 +147,55 @@ fn ingest(store: &SuggestStore, reingest: bool, providers: Vec<SuggestionProvide
         }
     };
 
-    store
+    let metrics = store
         .ingest(constraints)
         .unwrap_or_else(|e| panic!("Error in ingest: {e}"));
-    println!("Done");
+    if verbose {
+        print_header("Ingestion times");
+        let mut ingestion_times = metrics.ingestion_times;
+        let download_times: HashMap<String, u64> = metrics
+            .download_times
+            .into_iter()
+            .map(|s| (s.label, s.value))
+            .collect();
+
+        ingestion_times.sort_by_key(|s| s.value);
+        ingestion_times.reverse();
+        for sample in ingestion_times {
+            let label = &sample.label;
+            let ingestion_time = sample.value / 1000;
+            let download_time = download_times.get(label).unwrap_or(&0) / 1000;
+
+            println!(
+                "{label:30} Download: {download_time:>5}ms    Ingestion: {ingestion_time:>5}ms"
+            );
+        }
+    }
+    print_header("Done");
 }
 
-fn query(store: &SuggestStore, provider: SuggestionProviderArg, input: String) {
+fn query(
+    store: &SuggestStore,
+    provider: Option<SuggestionProviderArg>,
+    input: String,
+    verbose: bool,
+) {
     let query = SuggestionQuery {
-        providers: vec![provider.into()],
+        providers: match provider {
+            Some(provider) => vec![provider.into()],
+            None => SuggestionProvider::all().to_vec(),
+        },
         keyword: input,
         limit: None,
     };
-    let suggestions = store
-        .query(query)
+    let mut results = store
+        .query_with_metrics(query)
         .unwrap_or_else(|e| panic!("Error querying store: {e}"));
-    if suggestions.is_empty() {
-        println!("No Results");
+    if results.suggestions.is_empty() {
+        print_header("No Results");
     } else {
-        println!("Results:");
-        for suggestion in suggestions {
+        print_header("Results");
+        for suggestion in results.suggestions {
             let title = suggestion.title();
             let url = suggestion.url().unwrap_or("[no-url]");
             let icon = if suggestion.icon_data().is_some() {
@@ -171,4 +206,22 @@ fn query(store: &SuggestStore, provider: SuggestionProviderArg, input: String) {
             println!("{title} ({url}) ({icon})");
         }
     }
+    if verbose {
+        print_header("Query times");
+        results.query_times.sort_by_key(|s| s.value);
+        results.query_times.reverse();
+        for s in results.query_times {
+            println!("{:33} Time: {:>5}us", s.label, s.value);
+        }
+    }
+}
+
+fn print_header(msg: impl Into<String>) {
+    let mut msg = msg.into();
+    if msg.len() % 2 == 1 {
+        msg.push(' ');
+    }
+    let width = (70 - msg.len() - 2) / 2;
+    println!();
+    println!("{} {msg} {}", "=".repeat(width), "=".repeat(width));
 }